Script to update annotations
The script takes the pathname of a CSV or XLSX (Microsoft Excel
spreadsheet) file on the command line and updates the annotations in
the database from it.  This affects only the `taxon_annotation` table.
Old annotations that are updated will be labelled as "old" and newly
inserted annotations will be labelled as "valid".  Sequences in the
provided file taht are *not* found in the database will be flagged for
the user, and skipped.

Current status: No update is done, but ASV IDs for the given sequences
are verified against the sequences available in the database.