Restore ID concatenations in IPT views and explain in guide
Ideally, we want to output globally unique IDs for events and 
occurrences in DwC, 
but do not know whether users will provide such id:s. We now 
concatenate datasetID with eventID, by default, and explain this in our 
template. This may prompt users to only apply dataset-unique eventIDs, 
but if not, we also strip datasetID from eventID in R processing.